Add /power command
Replaces the /on, /off, /switchon, and /switchoff commands with a unified /power command.

The new /power command displays all plugs connected to OctoPrint, even if they are managed by different power manager plugins. It allows users to check the current state of each plug and toggle power on or off from a single interface.

This enables control of multiple devices - not just the printer's power supply, but also other peripherals such as heaters, LEDs, etc. - and lays the foundation for future support of GPIO and other power management systems (see #445).